Last Version

+++ ShippingUI/src/app/components/Trader-View/order-display-Trader/order-display-Trader.component.ts
@@ -6,6 +6,7 @@ import { OrderService } from 'src/app/Core/Services/order.service';
 import { AuthService } from 'src/app/Core/Services/auth.service';
 import { AddOrderComponent } from '../add-order/add-order.component';
 import { EditOrderComponent } from '../edit-order/edit-order.component';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-order-display-Trader',
@@ -62,11 +63,43 @@ export class OrderDispalyTraderComponent implements OnInit {
   }
 
   deleteOrder(orderId: number) {
-    this.orderService
-      .deleteOrderForTrader(orderId)
-      .subscribe((response: any) => {
-        this.loadOrders();
-      });
+    Swal.fire({
+      title: 'Are you sure you would like to delete this order?',
+      icon: 'warning',
+      iconColor: '#FFC700',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, cancel it!',
+      confirmButtonColor: '#00b2ff',
+      cancelButtonText: 'No, return',
+      width: '416px',
+      cancelButtonColor: '#eff2f5',
+    }).then((result) => {
+      if (result.value) {
+        this.orderService
+          .deleteOrderForTrader(orderId)
+          .subscribe((response: any) => {
+            Swal.fire({
+              title: 'Order has been successfully Deleted!',
+              icon: 'success',
+              confirmButtonColor: '#00b2ff',
+            });
+          });
+      } else {
+        Swal.fire({
+          title: 'Your form has not been cancelled!.',
+          icon: 'error',
+          confirmButtonText: 'Ok, got it!',
+          confirmButtonColor: '#00b2ff',
+          width: '416px',
+          iconColor: '#F1416C',
+          customClass: {
+            icon: 'custom-cancel-icon',
+            title: 'custom-content-class',
+          },
+        });
+      }
+    });
+    this.loadOrders();
   }
 
   filterData(inputValue: string) {
